完成 trace 功能（依旧存在部分疑问 就是 获取数据|=啥意思 为啥不直接=），新建一个映射map（syscall 135行是啥意思 以及为啥要空一行）

+++ kernel/proc.h
@@ -104,5 +104,5 @@ struct proc {
   struct inode *cwd;           // Current directory
   char name[16];               // Process name (debugging)
 
-  int tracemask;               //用来进行 筛选 trace函数的输出内容的
+  uint64 tracemask;               //用来进行 筛选 trace函数的输出内容的
 };
